requirements.txt
Django==3.1
psycopg2-binary
djangorestframework==3.12
djangorestframework-jsonapi==3.2.0
drf-spectacular
django-filter
django_q
redis
sentry-sdk
ptvsd